Seperated WallTileFactory to own files
