+++ src/TileSet.cpp
@@ -23,6 +23,8 @@
 #include "RenderModel.hpp"
 #include "tiled-map-loader.hpp"
 
+#include "WallTileFactory.hpp"
+
 // can always clean up embedded testing later
 #include <common/TestSuite.hpp>
 
@@ -32,492 +34,9 @@ namespace {
 
 using namespace cul::exceptions_abbr;
 
-using ConstTileSetPtr    = TileSet::ConstTileSetPtr;
-using TileSetPtr         = TileSet::TileSetPtr;
-using Triangle           = TriangleSegment;
-#if 0
-using CardinalDirections = TileFactory::CardinalDirections;
-#endif
-class TranslatableTileFactory : public TileFactory {
-public:
-    void setup(Vector2I, const tinyxml2::XMLElement * properties, Platform::ForLoaders &) override;
-
-protected:
-    Vector translation() const { return m_translation; }
-
-    Entity make_entity(Platform::ForLoaders & platform, Vector2I tile_loc,
-                       SharedCPtr<RenderModel> model_ptr) const;
-
-private:
-    Vector m_translation;
-};
-
-CardinalDirections cardinal_direction_from(const char * str) {
-    auto seq = [str](const char * s) { return !::strcmp(str, s); };
-    using Cd = CardinalDirections;
-    if (seq("n" )) return Cd::n;
-    if (seq("s" )) return Cd::s;
-    if (seq("e" )) return Cd::e;
-    if (seq("w" )) return Cd::w;
-    if (seq("ne")) return Cd::ne;
-    if (seq("nw")) return Cd::nw;
-    if (seq("se")) return Cd::se;
-    if (seq("sw")) return Cd::sw;
-    throw InvArg{""};
-}
-
-// ------------------------------- <! messy !> --------------------------------
-#if 0
-bool is_solution(Real);
-#endif
-
-auto make_get_next_for_dir_split_v(Vector end, Vector step) {
-    return [end, step] (Vector east_itr) {
-        auto cand_next = east_itr + step;
-        if (are_very_close(cand_next, end)) return cand_next;
-
-        if (are_very_close(normalize(end - east_itr ),
-                           normalize(end - cand_next)))
-        { return cand_next; }
-        return end;
-    };
-};
-
-template <typename Func>
-void make_linear_triangle_strip
-    (const Vector & a_start, const Vector & a_last,
-     const Vector & b_start, const Vector & b_last,
-     Real step, Func && f)
-{
-    const auto make_step = [step](const Vector & start, const Vector & last) {
-        auto diff = last - start;
-        if (are_very_close(diff, Vector{})) return Vector{};
-        return step*normalize(diff);
-    };
-
-    auto itr_a = a_start;
-    const auto next_a = make_get_next_for_dir_split_v(
-        a_last, make_step(a_start, a_last));
-
-    auto itr_b = b_start;
-    const auto next_b = make_get_next_for_dir_split_v(
-        b_last, make_step(b_start, b_last));
-
-    while (   !are_very_close(itr_a, a_last)
-           && !are_very_close(itr_b, b_last))
-    {
-        const auto new_a = next_a(itr_a);
-        const auto new_b = next_b(itr_b);
-        f(Triangle{itr_a, itr_b, new_a});
-        f(Triangle{itr_a, new_a, new_b});
-        itr_a = new_a;
-        itr_b = new_b;
-    }
-
-    if (!are_very_close(itr_a, a_last)) {
-        f(Triangle{itr_a, b_last, a_last});
-    } else if (!are_very_close(itr_b, b_last)) {
-        f(Triangle{itr_b, itr_a, b_last});
-    }
-}
-
-enum SplitOpt {
-    k_flats_only = 1 << 0,
-    k_wall_only  = 1 << 1,
-    k_both_flats_and_wall = k_flats_only | k_wall_only
-};
-
-// should handle division == 0, == 1
-// everything around
-// {-0.5, x,  0.5}, {0.5, x,  0.5}
-// {-0.5, x, -0.5}, {0.5, x, -0.5}
-template <SplitOpt k_opt, typename Func>
-void north_south_split
-    (Real north_east_y, Real north_west_y,
-     Real south_east_y, Real south_west_y,
-     Real division_z, Func && f);
-
-template <SplitOpt k_opt, typename Func>
-void east_west_split
-    (Real north_east_y, Real north_west_y,
-     Real south_east_y, Real south_west_y,
-     Real division_z, Func && f);
-
-auto make_get_next_for_dir_split(Vector end, Real dir) {
-    return [end, dir] (Vector east_itr) {
-        auto cand_next = east_itr + dir*Vector{0, 1, 0};
-        if (are_very_close(cand_next, end)) return cand_next;
-
-        if (are_very_close(normalize(end - east_itr ),
-                           normalize(end - cand_next)))
-        { return cand_next; }
-        return end;
-    };
-};
-
-
-template <SplitOpt k_opt, typename Func>
-void north_south_split
-    (Real north_east_y, Real north_west_y,
-     Real south_east_y, Real south_west_y,
-     Real division_z, Func && f)
-{
-    // it really doesn't have to be "elegant"
-    // division z must make sense
-    assert(division_z >= -0.5 && division_z <= 0.5);
-
-    // both sets of y values' directions must be the same
-    assert((north_east_y - north_west_y)*(south_east_y - south_west_y) >= 0);
-
-    const Vector div_nw{-0.5, north_west_y, division_z};
-    const Vector div_ne{ 0.5, north_east_y, division_z};
-
-    const Vector div_sw{-0.5, south_west_y, division_z};
-    const Vector div_se{ 0.5, south_east_y, division_z};
-    // We must handle division_z being 0.5
-    if constexpr ((k_opt & k_flats_only) && false) {
-#       if 0
-        if (!are_very_close(division_z, 0.5)) {
-            Vector nw{-0.5, north_west_y, 0.5};
-            Vector ne{ 0.5, north_east_y, 0.5};
-            f(Triangle{nw, ne    , div_nw});
-            f(Triangle{nw, div_nw, div_ne});
-        }
-        if (!are_very_close(division_z, -0.5)) {
-            Vector sw{-0.5, south_west_y, 0.5};
-            Vector se{ 0.5, south_east_y, 0.5};
-            f(Triangle{sw, se    , div_sw});
-            f(Triangle{sw, div_sw, div_se});
-        }
-#       endif
-        Vector nw{-0.5, north_west_y, 0.5};
-        Vector ne{ 0.5, north_east_y, 0.5};
-        make_linear_triangle_strip(nw, div_nw, ne, div_ne, 1, f);
-        Vector sw{-0.5, south_west_y, 0.5};
-        Vector se{ 0.5, south_east_y, 0.5};
-        make_linear_triangle_strip(div_sw, sw, div_se, se, 1, f);
-    }
-    // We should only skip triangles along the wall if
-    // there's no elevation difference to cover
-    if constexpr (k_opt & k_wall_only) {
-        make_linear_triangle_strip(div_nw, div_sw, div_ne, div_se, 1, f);
-#       if 0
-        // it doesn't matter so much where I start, so long as
-        // 1ue steps are used
-        auto normalize_or_zero = [] (Real x) { return are_very_close(x, 0) ? 0 : normalize(x); };
-        // must accept that one direction or both maybe "0"
-        const auto east_dir = normalize_or_zero(north_east_y - north_west_y);
-        const auto west_dir = normalize_or_zero(south_east_y - south_west_y);
-        if (are_very_close(east_dir, 0) && are_very_close(west_dir, 0))
-            return;
-
-        auto east_itr = div_ne;
-        const auto & east_end = div_se;
-
-        auto west_itr = div_nw;
-        const auto & west_end = div_sw;
-
-        auto get_east_next = make_get_next_for_dir_split(east_end, east_dir);
-        auto get_west_next = make_get_next_for_dir_split(west_end, west_dir);
-
-        while (   !are_very_close(east_itr, east_end)
-               || !are_very_close(west_itr, west_end))
-        {
-            auto east_next = get_east_next(east_itr);
-            auto west_next = get_west_next(west_itr);
-
-            if (!are_very_close(east_itr, east_end)) {
-                f(Triangle{east_itr, west_itr, east_next});
-            }
-            if (!are_very_close(west_itr, west_end)) {
-                f(Triangle{west_itr, east_itr, west_next});
-            }
-
-            east_itr = east_next;
-            west_itr = west_next;
-        }
-#   endif
-    }
-}
-
-template <SplitOpt k_opt, typename Func>
-void east_west_split
-    (Real east_north_y, Real east_south_y,
-     Real west_north_y, Real west_south_y,
-     Real division_x, Func && f)
-{
-    // simply switch roles
-    // east <-> north
-    // west <-> south
-    auto remap_vector = [] (const Vector & r) { return Vector{r.z, r.y, r.x}; };
-    north_south_split<k_opt>(
-        east_north_y, east_south_y, west_north_y, west_south_y,
-        division_x,
-        [f = std::move(f), remap_vector] (const Triangle & tri)
-    {
-        f(Triangle{
-            remap_vector(tri.point_a()),
-            remap_vector(tri.point_b()),
-            remap_vector(tri.point_c())
-        });
-    });
-}
-
-class WallTileFactory final : public TranslatableTileFactory {
-public:
-    void assign_render_model_wall_cache(WallRenderModelCache & cache) final
-        { m_render_model_cache = &cache; }
-
-    static WallElevationAndDirection elevations_and_direction
-        (const NeighborInfo & ninfo, Real known_elevation,
-         CardinalDirections dir, Vector2I tile_loc)
-    {
-        WallElevationAndDirection rv;
-        // I need a way to address each corner...
-        using Cd = CardinalDirections;
-        rv.direction = dir;
-        rv.tileset_location = tile_loc;
-        auto known_corners = make_known_corners(rv.direction);
-        for (auto corner : { Cd::nw, Cd::sw, Cd::se, Cd::ne }) {
-            // walls are only generated for dips on "unknown corners"
-            // if a neighbor elevation is unknown, then no wall it created for that
-            // corner (which can very easily mean no wall are generated on any "dip"
-            // corner
-            Real neighbor_elevation = ninfo.neighbor_elevation(corner);
-            Real diff   = known_elevation - neighbor_elevation;
-            bool is_dip =    cul::is_real(neighbor_elevation)
-                          && known_elevation > neighbor_elevation
-                          && !known_corners[corner_index(corner)];
-            // must be finite for our purposes
-            rv.dip_heights[corner_index(corner)] = is_dip ? diff : 0;
-        }
-
-        return rv;
-    }
-
-    static int corner_index(CardinalDirections dir) {
-        using Cd = CardinalDirections;
-        switch (dir) {
-        case Cd::nw: return 0;
-        case Cd::sw: return 1;
-        case Cd::se: return 2;
-        case Cd::ne: return 3;
-        default: break;
-        }
-        throw InvArg{""};
-    }
-
-private:
-    static constexpr const Real k_visual_dip_thershold = 0.5;
-    static constexpr const Real k_physical_dip_thershold = 1;
-
-    template <typename Iter>
-    static void translate_points(Vector r, Iter beg, Iter end) {
-        for (auto itr = beg; itr != end; ++itr) {
-            *itr += r;
-        }
-    }
-
-    template <typename Iter>
-    static void scale_points_x(Real x, Iter beg, Iter end) {
-        for (auto itr = beg; itr != end; ++itr) {
-            itr->x *= x;
-        }
-    }
-
-    template <typename Iter>
-    static void scale_points_z(Real x, Iter beg, Iter end) {
-        for (auto itr = beg; itr != end; ++itr) {
-            itr->z *= x;
-        }
-    }
-
-    void setup
-        (Vector2I loc_in_ts, const tinyxml2::XMLElement * properties, Platform::ForLoaders & platform) final
-    {
-        TranslatableTileFactory::setup(loc_in_ts, properties, platform);
-        m_dir = cardinal_direction_from(find_property("direction", properties));
-        m_tileset_location = loc_in_ts;
-    }
-
-    Slopes tile_elevations() const final {
-        // it is possible that some elevations are indeterminent...
-        Real y = translation().y + 1;
-        return Slopes{0, y, y, y, y};
-    }
-
-    // a wall has at least one elevation that's known
-    // a wall only generates if it's dip sides connect to tiles whose elevations
-    // are lower than the known elevation
-
-    void operator ()
-        (EntityAndTrianglesAdder & adder, const NeighborInfo & ninfo,
-         Platform::ForLoaders & platform) const final
-    { make_tile(adder, ninfo, platform); }
-
-    void make_tile
-        (EntityAndTrianglesAdder & adder, const NeighborInfo & ninfo,
-         Platform::ForLoaders & platform) const
-    {
-        auto wed = elevations_and_direction(ninfo);
-        if (m_render_model_cache) {
-            auto itr = m_render_model_cache->find(wed);
-            if (itr != m_render_model_cache->end()) {
-                const auto & [render_model, triangles] = itr->second; {}
-                return make_entities_and_triangles(adder, platform, ninfo, render_model, triangles);
-            }
-        }
-        auto [render_model, triangles] = make_render_model_and_triangles(wed, ninfo, platform); {}
-        if (m_render_model_cache) {
-            m_render_model_cache->insert(std::make_pair(
-                wed,
-                make_tuple(std::move(render_model), std::move(triangles))
-            ));
-            assert(m_render_model_cache->find(wed) != m_render_model_cache->end());
-            return make_tile(adder, ninfo, platform);
-        }
-        // in case there's no cache...
-        make_entities_and_triangles(adder, platform, ninfo, render_model, triangles);
-    }
-
-    void make_entities_and_triangles(
-        EntityAndTrianglesAdder & adder,
-        Platform::ForLoaders & platform,
-        const NeighborInfo & ninfo,
-        const SharedPtr<const RenderModel> & render_model,
-        const std::vector<Triangle> & triangles) const
-    {
-        for (auto & triangle : triangles) {
-#           if 0
-            // | there's *got* to be a way to generate this stuff using a
-            // v "shared" deletor
-
-            auto triangle_vec = new std::vector<Triangle>{triangles};
-            auto del = [triangle_vec] (Triangle *) { delete triangle_vec; };
-            for (auto & triangle : *triangle_vec) {
-                SharedPtr<const Triangle>{&triangle, del};
-            }
-            //SharedPtr
-#           endif
-            adder.add_triangle(Triangle{triangle.move(translation())});
-        }
-        // mmm
-        // make_entity(platform, ninfo.tile_location(), render_model);
-    }
-
-    static std::array<bool, 4> make_known_corners(CardinalDirections dir) {
-        using Cd = CardinalDirections;
-        auto mk_rv = [](bool nw, bool sw, bool se, bool ne) {
-            std::array<bool, 4> rv;
-            const std::array k_corners = {
-                make_tuple(Cd::nw, nw), make_tuple(Cd::ne, ne),
-                make_tuple(Cd::sw, sw), make_tuple(Cd::se, se),
-            };
-            for (auto [corner, val] : k_corners)
-                rv[corner_index(corner)] = !val;
-            return rv;
-        };
-        switch (dir) {
-        // a north wall, all point on the north are known
-        case Cd::n : return mk_rv(true , false, false, true );
-        case Cd::s : return mk_rv(false, true , true , false);
-        case Cd::e : return mk_rv(false, false, true , true );
-        case Cd::w : return mk_rv(true , true , false, false);
-        case Cd::nw: return mk_rv(true , false, false, false);
-        case Cd::sw: return mk_rv(false, true , false, false);
-        case Cd::se: return mk_rv(false, false, true , false);
-        case Cd::ne: return mk_rv(false, false, false, true );
-        default: break;
-        }
-        throw BadBranchException{__LINE__, __FILE__};
-    }
-
-    WallElevationAndDirection elevations_and_direction
-        (const NeighborInfo & ninfo) const
-    {
-        return WallTileFactory::elevations_and_direction
-            (ninfo, translation().y + 1, m_dir, m_tileset_location);
-    }
-
-    Tuple<SharedPtr<const RenderModel>, std::vector<Triangle>>
-        make_render_model_and_triangles(
-        const WallElevationAndDirection & wed,
-        const NeighborInfo & ninfo,
-        Platform::ForLoaders &) const
-    {
-        auto offset = grid_position_to_v3(ninfo.tile_location())
-            + translation() + Vector{0, 1, 0};
-        std::vector<Triangle> triangles;
-        struct Impl final {
-            Impl(std::vector<Triangle> & triangles_):
-                triangles(triangles_)
-            {}
-
-            void operator () (const Triangle & triangle) const {
-                triangles.push_back(triangle);
-            }
-
-            std::vector<Triangle> & triangles;
-        };
-        Impl add_triangle_{triangles};
-        auto & add_triangle = add_triangle_;
-#       if 0
-        auto add_triangle = [&triangles] (const Triangle & triangle) {
-            triangles.push_back(triangle);
-        };
-#       endif
-        // we have cases depending on the number of dips
-        // okay, I should like to handle only two (adjacent) and three dips
-        using Cd = CardinalDirections;
-        auto adjusted_thershold = k_physical_dip_thershold - 0.5;
-        // nvm, direction describes which case
-        switch (m_dir) {
-        // a north wall, all point on the north are known
-        case Cd::n :
-            north_south_split<k_both_flats_and_wall>(
-                offset.y - wed.dip_heights[corner_index(Cd::ne)], // ne
-                offset.y - wed.dip_heights[corner_index(Cd::nw)],
-                offset.y, offset.y,
-                adjusted_thershold, add_triangle);
-            break;
-        case Cd::s :
-            north_south_split<k_both_flats_and_wall>(
-                offset.y, offset.y,
-                offset.y - wed.dip_heights[corner_index(Cd::se)], // se
-                offset.y - wed.dip_heights[corner_index(Cd::sw)],
-                1 - adjusted_thershold, add_triangle);
-            break;
-        case Cd::e :
-            east_west_split<k_both_flats_and_wall>(
-                offset.y - wed.dip_heights[corner_index(Cd::ne)], // ne
-                offset.y - wed.dip_heights[corner_index(Cd::se)],
-                offset.y, offset.y,
-                adjusted_thershold, add_triangle);
-            break;
-        case Cd::w :
-            east_west_split<k_both_flats_and_wall>(
-                offset.y, offset.y,
-                offset.y - wed.dip_heights[corner_index(Cd::nw)], // ne
-                offset.y - wed.dip_heights[corner_index(Cd::sw)],
-                1 - adjusted_thershold, add_triangle);
-            break;
-        // maybe have seperate divider functions for these cases?
-        case Cd::nw: ;
-        case Cd::sw: ;
-        case Cd::se: ;
-        case Cd::ne: ;
-            throw RtError{"unimplemented"};
-        default: break;
-        }
-        return make_tuple(nullptr, std::move(triangles));
-    }
-
-    CardinalDirections m_dir = CardinalDirections::ne;
-    WallRenderModelCache * m_render_model_cache = nullptr;
-    Vector2I m_tileset_location;
-
-    // I still need to known the wall texture coords
-};
+using ConstTileSetPtr = TileSet::ConstTileSetPtr;
+using TileSetPtr      = TileSet::TileSetPtr;
+using Triangle        = TriangleSegment;
 
 // ----------------------------------------------------------------------------
 
@@ -848,147 +367,6 @@ Real TileFactory::NeighborInfo::neighbor_elevation(CardinalDirections dir) const
     return (*itr->second)();
 }
 
-class TestTrianglesAdder final : public EntityAndTrianglesAdder {
-public:
-    void add_triangle(const TriangleSegment & triangle) final
-        { triangles.push_back(triangle); }
-
-    void add_entity(const Entity &) final {}
-
-    std::vector<Triangle> triangles;
-};
-
-/* static */ bool TileFactory::test_wall_factory() {
-#   define mark MACRO_MARK_POSITION_OF_CUL_TEST_SUITE
-    using namespace cul::ts;
-    TestSuite suite;
-
-    constexpr const auto k_tileset_fn = "test-tileset.tsx";
-    static auto load_tileset = [](const char * fn, TileSet & tileset) {
-        TiXmlDocument document;
-        document.LoadFile(fn);
-        auto & platform = Platform::null_callbacks();
-        tileset.load_information(platform, *document.RootElement());
-    };
-    static constexpr const int k_connecting_tile = 16;
-    static constexpr const int k_north_wall_no_translation = 34;
-    static const auto k_sample_layer = [] {
-        Grid<int> layer;
-        layer.set_size(1, 2);
-        layer(0, 0) = k_connecting_tile;
-        layer(0, 1) = k_north_wall_no_translation;
-        return layer;
-    } ();
-    static const auto make_sample_neighbor_info = [] (const TileSet & tileset) {
-        return NeighborInfo{tileset, k_sample_layer, Vector2I{0, 1}, Vector2I{}};
-    };
-    static const auto verify_tile_factory = [] (const TileFactory * fact) {
-        if (fact) return fact;
-        throw RtError{"Uh Oh, no tile factory"};
-    };
-
-    suite.start_series("TileFactory::NeighborInfo");
-    // no neighbor, no real number is returned
-    mark(suite).test([] {
-        auto res = NeighborInfo::make_no_neighbor()
-            .neighbor_elevation(CardinalDirections::nw);
-        return test(!cul::is_real(res));
-    });
-    // yes neighbor, return real number
-    mark(suite).test([] {
-        TileSet tileset;
-        load_tileset(k_tileset_fn, tileset);
-        auto ninfo = make_sample_neighbor_info(tileset);
-        auto res = ninfo.neighbor_elevation(CardinalDirections::nw);
-        return test(cul::is_real(res));
-    });
-    // against another wall, no real return
-    mark(suite).test([] {
-        TileSet tileset;
-        load_tileset(k_tileset_fn, tileset);
-        Grid<int> layer;
-        layer.set_size(1, 2);
-        layer(0, 0) = k_north_wall_no_translation;
-        layer(0, 1) = k_north_wall_no_translation;
-        NeighborInfo ninfo{tileset, layer, Vector2I{0, 1}, Vector2I{}};
-        auto res = ninfo.neighbor_elevation(CardinalDirections::nw);
-        return test(cul::is_real(res));
-    });
-
-    suite.start_series("Wall Tile Factory");
-    mark(suite).test([] {
-        // not possible without neighbors!
-        TileSet tileset;
-        load_tileset(k_tileset_fn, tileset);
-        using Wtf = WallTileFactory;
-        using Cd = CardinalDirections;
-
-        auto wed = Wtf::elevations_and_direction
-            (make_sample_neighbor_info(tileset), 1, Cd::n, Vector2I{});
-        // both dip heights should be 2 in this case
-        auto nw = wed.dip_heights[Wtf::corner_index(Cd::nw)];
-        auto ne = wed.dip_heights[Wtf::corner_index(Cd::ne)];
-        auto sw = wed.dip_heights[Wtf::corner_index(Cd::sw)];
-        auto se = wed.dip_heights[Wtf::corner_index(Cd::se)];
-        return test(   are_very_close(ne, 1) && are_very_close(nw, 1)
-                    && are_very_close(sw, 0) && are_very_close(se, 0));
-    });
-
-    // if there are no neighbors, then no walls should be generated!
-    mark(suite).test([] {
-        using Wtf = WallTileFactory;
-        using Cd = CardinalDirections;
-
-        auto wed = Wtf::elevations_and_direction
-            (NeighborInfo::make_no_neighbor(), 1, Cd::n, Vector2I{});
-        auto nw = wed.dip_heights[Wtf::corner_index(Cd::nw)];
-        auto ne = wed.dip_heights[Wtf::corner_index(Cd::ne)];
-        auto sw = wed.dip_heights[Wtf::corner_index(Cd::sw)];
-        auto se = wed.dip_heights[Wtf::corner_index(Cd::se)];
-        return test(   are_very_close(ne, 0) && are_very_close(nw, 0)
-                    && are_very_close(sw, 0) && are_very_close(se, 0));
-    });
-    mark(suite).test([] {
-        TileSet tileset;
-        load_tileset(k_tileset_fn, tileset);
-        auto * wall_tile_factory = verify_tile_factory(tileset(k_north_wall_no_translation));
-
-        TestTrianglesAdder adder;
-        // what the tile is, appears through the adder
-
-        auto sample_neighbor = make_sample_neighbor_info(tileset);
-        (*wall_tile_factory)(adder, sample_neighbor, Platform::null_callbacks());
-        bool wall_found = std::any_of(adder.triangles.begin(), adder.triangles.end(),
-            [](const Triangle & tptr)
-        {
-            return    are_very_close(tptr.point_a().z, tptr.point_b().z)
-                   && are_very_close(tptr.point_b().z, tptr.point_c().z);
-        });
-        return test(wall_found);
-    });
-    // it *is* correct behavior to not generate walls when there is no neighbor
-    mark(suite).test([] {
-        TileSet tileset;
-        load_tileset(k_tileset_fn, tileset);
-        auto * wall_tile_factory = verify_tile_factory(tileset(k_north_wall_no_translation));
-
-        TestTrianglesAdder adder;
-        // what the tile is, appears through the adder
-
-        auto sample_neighbor = NeighborInfo::make_no_neighbor();
-        (*wall_tile_factory)(adder, sample_neighbor, Platform::null_callbacks());
-        bool wall_not_found = std::none_of(adder.triangles.begin(), adder.triangles.end(),
-            [](const Triangle & tptr)
-        {
-            return    are_very_close(tptr.point_a().z, tptr.point_b().z)
-                   && are_very_close(tptr.point_b().z, tptr.point_c().z);
-        });
-        return test(wall_not_found);
-    });
-#   undef mark
-    return suite.has_successes_only();
-}
-
 void TileFactory::set_shared_texture_information
     (const SharedCPtr<Texture> & texture_ptr_, const Size2 & texture_size_,
      const Size2 & tile_size_)
@@ -1074,35 +452,6 @@ void TileFactory::set_shared_texture_information
 }
 
 namespace { // ----------------------------------------------------------------
-
-void TranslatableTileFactory::setup
-    (Vector2I, const tinyxml2::XMLElement * properties, Platform::ForLoaders &)
-{
-    // eugh... having to run through elements at a time
-    // not gonna worry about it this iteration
-    if (const auto * val = find_property("translation", properties)) {
-        auto list = { &m_translation.x, &m_translation.y, &m_translation.z };
-        auto itr = list.begin();
-        for (auto value_str : split_range(val, val + ::strlen(val),
-                                          is_comma, make_trim_whitespace<const char *>()))
-        {
-            bool is_num = cul::string_to_number(value_str.begin(), value_str.end(), **itr);
-            assert(is_num);
-            ++itr;
-        }
-    }
-}
-
-/* protected */ Entity
-    TranslatableTileFactory::make_entity
-    (Platform::ForLoaders & platform, Vector2I tile_loc,
-     SharedCPtr<RenderModel> model_ptr) const
-{
-    return TileFactory::make_entity(platform,
-        m_translation + grid_position_to_v3(tile_loc), model_ptr);
-}
-
-// --------------------------- <anonymous> namespace --------------------------
 
 void SlopesBasedModelTile::operator ()
     (EntityAndTrianglesAdder & adder, const NeighborInfo & ninfo,
