+++ src/TriangleSegment.cpp
@@ -33,19 +33,13 @@ using LimitIntersection = TriangleSegment::LimitIntersection;
 using cul::find_smallest_diff, cul::make_nonsolution_sentinel,
       cul::make_zero_vector;
 using std::min_element;
-#if 0
-template <typename T>
-using EnableBoolIfVec = std::enable_if_t<cul::k_is_vector_type<T>, bool>;
-#endif
+
 Real find_intersecting_position_for_first
     (Vector2 first_line_a , Vector2 first_line_b ,
      Vector2 second_line_a, Vector2 second_line_b);
 
 bool is_solution(Real x);
-#if 0
-template <typename Vec>
-EnableBoolIfVec<Vec> are_parallel(const Vec & a, const Vec & b);
-#endif
+
 Real get_component_for_basis(const Vector & pt_on_place, const Vector & basis);
 
 Vector2 find_point_c_in_2d(const Vector & a, const Vector & b, const Vector & c);
@@ -93,6 +87,9 @@ TriangleSegment::TriangleSegment
     check_invarients();
 }
 
+Real TriangleSegment::area_of_triangle() const
+    { return cul::area_of_triangle(point_a(), point_b(), point_c()); }
+
 Vector TriangleSegment::basis_i() const
     { return normalize(point_b() - point_a()); }
 
@@ -102,6 +99,11 @@ Vector TriangleSegment::basis_j() const {
     auto rv = cross(normal(), basis_i());
     assert(are_very_close(magnitude(rv), 1.));
     return rv;
+}
+
+bool TriangleSegment::can_be_projected_onto(const Vector & n) const noexcept {
+    auto [a, b, c] = project_onto_plane_(n);
+    return !are_parallel(b - a, b - c);
 }
 
 Vector TriangleSegment::center() const noexcept
@@ -148,7 +150,7 @@ SideCrossing TriangleSegment::check_for_side_crossing
         old + new_);
 }
 
-Vector2 TriangleSegment::closest_contained_point(Vector p) const {
+Vector2 TriangleSegment::closest_contained_point(const Vector & p) const {
     check_invarients();
     auto r = closest_point(p);
     // do something to r...
@@ -158,7 +160,7 @@ Vector2 TriangleSegment::closest_contained_point(Vector p) const {
     return check_for_side_crossing(center_in_2d(), r).inside;
 }
 
-Vector2 TriangleSegment::closest_point(Vector p) const {
+Vector2 TriangleSegment::closest_point(const Vector & p) const {
     check_invarients();
     // find via projection
     // https://math.stackexchange.com/questions/633181/formula-to-project-a-vector-onto-a-plane
@@ -170,7 +172,7 @@ Vector2 TriangleSegment::closest_point(Vector p) const {
                    get_component_for_basis(pa_on_plane, basis_j())};
 }
 
-bool TriangleSegment::contains_point(Vector2 r) const noexcept
+bool TriangleSegment::contains_point(const Vector2 & r) const noexcept
     { return is_real(r) ? point_region(r) == k_inside : false; }
 
 TriangleSegment TriangleSegment::flip() const noexcept {
@@ -181,34 +183,8 @@ TriangleSegment TriangleSegment::flip() const noexcept {
     return rv;
 }
 
-Vector2 TriangleSegment::intersection(Vector a, Vector b) const {
-    return limit_with_intersection(a, b).intersection;
-#   if 0
-    check_invarients();
-    // multiple sources on this
-    // from stackoverflow (multiple posts), blender, rosetta code, and on, and on
-    // again, maintaining that simple geometic formulae as uncopyrightable
-    constexpr const auto k_no_sol = make_nonsolution_sentinel<Vector2>();
-
-    auto norm = normal();
-    auto ba = b - a;
-    auto denom = dot(norm, ba);
-    if (are_very_close(denom, 0)) return k_no_sol;
-
-    auto back_from_head = dot(norm, b - point_a()) / denom;
-    // 0 at head, 1 at tail
-    if (back_from_head < 0 || back_from_head > 1)
-        { return k_no_sol; }
-
-    // [NTS] back_from_head > 1 will be before the intersection
-    //       I will need to find a value s.t. it's not counted as being on the
-    //       plane (that is intersecting)
-    auto on_plane = b - ba*back_from_head;
-    auto r = closest_point(on_plane);
-    // it's possible to hit the plane, but not be inside the triangle segment
-    return contains_point(r) ? r : k_no_sol;
-#   endif
-}
+Vector2 TriangleSegment::intersection(const Vector & a, const Vector & b) const
+    { return limit_with_intersection(a, b).intersection; }
 
 LimitIntersection TriangleSegment::limit_with_intersection
     (const Vector & a, const Vector & b) const
@@ -283,12 +259,17 @@ Vector TriangleSegment::point_c() const noexcept { return m_c; }
 Vector2 TriangleSegment::point_c_in_2d() const noexcept
     { return m_c_2d; }
 
-Vector TriangleSegment::point_at(Vector2 r) const {
+Vector TriangleSegment::point_at(const Vector2 & r) const {
     if (!is_real(r)) {
          throw InvArg{"TriangleSurface::point_at: given point must have all "
                       "real number components."};
     }
     return point_a() + basis_i()*r.x + basis_j()*r.y;
+}
+
+TriangleSegment TriangleSegment::project_onto_plane(const Vector & n) const {
+    auto [a, b, c] = project_onto_plane_(n); {}
+    return TriangleSegment{a, b, c};
 }
 
 Tuple<Vector, Vector> TriangleSegment::side_points(Side side) const {
@@ -314,7 +295,7 @@ Tuple<Vector2, Vector2> TriangleSegment::side_points_in_2d(Side side) const {
 }
 
 /* private */ void TriangleSegment::check_invarients() const noexcept {
-#   if 0
+#   ifdef MACRO_DEBUG
     assert(is_real(m_a));
     assert(is_real(m_b));
     assert(is_real(m_c));
@@ -364,10 +345,7 @@ Tuple<Vector2, Vector2> TriangleSegment::side_points_in_2d(Side side) const {
     auto center = center_in_2d();
     auto is_crossed_line = [r, center] (const Vector2 & a, const Vector2 & b)
         { return is_solution(find_intersecting_position_for_first(a, b, center, r)); };
-#   if 0
-    auto intersecting_pt = [center, r](const Vector2 & a, const Vector2 & b)
-        { return find_intersecting_position_for_first(a, b, center, r); };
-#   endif
+
     auto a = point_a_in_2d();
     auto b = point_b_in_2d();
     if (are_parallel(a - b, a - r)) return k_inside;
@@ -379,6 +357,14 @@ Tuple<Vector2, Vector2> TriangleSegment::side_points_in_2d(Side side) const {
     if (is_crossed_line(b, c)) return k_side_bc;
     if (is_crossed_line(c, a)) return k_side_ca;
     return k_inside;
+}
+
+/* private */ Tuple<Vector, Vector, Vector>
+    TriangleSegment::project_onto_plane_(const Vector & n) const noexcept
+{
+    return make_tuple(cul::project_onto_plane(point_a(), n),
+                      cul::project_onto_plane(point_b(), n),
+                      cul::project_onto_plane(point_c(), n));
 }
 
 const char * to_string(TriangleSide side) {
@@ -424,27 +410,7 @@ Real find_intersecting_position_for_first
 
 bool is_solution(Real x)
     { return !std::equal_to<Real>{}(x, k_no_intersection_2d); }
-#if 0
-template <typename Vec>
-EnableBoolIfVec<Vec> are_parallel(const Vec & a, const Vec & b) {
-#   if 0 // gdb *really* doesn't like me short circutting functions :c
-    auto mag = magnitude(normalize(a) + normalize(b));
-    return are_very_close(mag, 0) || are_very_close(mag, 2);
-#   elif 1
-    auto mk_zero = [] {
-        constexpr auto k_dim = cul::VectorTraits<Vec>::k_dimension_count;
-        static_assert(k_dim == 2 || k_dim == 3,
-            "<anonymous>::are_parallel: must be used only with 2 or 3 dimensions.");
-        if constexpr (k_dim == 2) return 0;
-        else return make_zero_vector<Vec>();
-    };
-    return are_very_close(mk_zero(), cross(a, b));
-#   else
-    auto frac = (dot(a, b)*dot(a, b)) / (sum_of_squares(a)*sum_of_squares(b));
-    return are_very_close(magnitude(frac), 1);
-#   endif
-}
-#endif
+
 Real get_component_for_basis(const Vector & pt_on_plane, const Vector & basis) {
     // basis is assumed to be a normal vector
     assert(are_very_close(magnitude(basis), 1.));
