+++ src/platform/test/test-main.cpp
@@ -24,14 +24,13 @@
 #include <algorithm>
 
 int main() {
-    using BoolFunc = bool(*)();
-
     // I *could* just dump all my test functions here I guess...
     const std::array k_test_functions = {
         run_triangle_segment_tests,
         run_map_loader_tests,
         run_triangle_links_tests,
-        run_systems_tests
+        run_systems_tests,
+        run_wall_tile_factory_tests
     };
     std::array<bool, k_test_functions.size()> results;
     for (std::size_t i = 0; i != k_test_functions.size(); ++i)
